# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A comprehensive API for hotel menu management, including hotels, categories, menu items, chefs, ratings, discounts, analytics, and feedback.",
        "title": "Hotel Menu API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/menu-items": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new menu item (admin/superadmin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Items"
                ],
                "summary": "Create menu item",
                "parameters": [
                    {
                        "description": "Menu item data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CreateMenuItemRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MenuItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/menu-items/{id}": {
            "get": {
                "description": "Get a menu item's details by its UUID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get menu item by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Menu Item ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MenuItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a menu item's details (admin/superadmin only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Items"
                ],
                "summary": "Update menu item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Menu Item ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Menu item update data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.UpdateMenuItemRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MenuItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a menu item (admin/superadmin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu Items"
                ],
                "summary": "Delete menu item",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Menu Item ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/menu/categories/{id}/menu-items": {
            "get": {
                "description": "Get all menu items for a specific category",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "List menu items by category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.MenuItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/menu/hotels/{id}/menu-items": {
            "get": {
                "description": "Get all menu items for a specific hotel",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "List menu items by hotel",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Hotel ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/domain.MenuItem"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
        "/menu/menu-items/slug/{slug}": {
            "get": {
                "description": "Get a menu item's details by its URL-friendly slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Menu"
                ],
                "summary": "Get menu item by slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Menu item slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/domain.MenuItem"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/utils.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CreateMenuItemRequest": {
            "type": "object",
            "required": [
                "category_id",
                "hotel_id",
                "name_en",
                "price"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "chef_id": {
                    "type": "string"
                },
                "description_am": {
                    "type": "string"
                },
                "description_en": {
                    "type": "string"
                },
                "hotel_id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "is_special": {
                    "type": "boolean"
                },
                "name_am": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "domain.MenuItem": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "chef_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description_am": {
                    "type": "string"
                },
                "description_en": {
                    "type": "string"
                },
                "hotel_id": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "is_special": {
                    "type": "boolean"
                },
                "menu_item_id": {
                    "type": "string"
                },
                "name_am": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "slug": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "video_url": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                }
            }
        },
...
        "domain.UpdateMenuItemRequest": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "chef_id": {
                    "type": "string"
                },
                "description_am": {
                    "type": "string"
                },
                "description_en": {
                    "type": "string"
                },
                "image_url": {
                    "type": "string"
                },
                "is_available": {
                    "type": "boolean"
                },
                "is_special": {
                    "type": "boolean"
                },
                "name_am": {
                    "type": "string"
                },
                "name_en": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "video_url": {
                    "type": "string"
                }
            }
        },
...
        "utils.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}